+++ tukaan/_typing.py
@@ -6,10 +6,7 @@ if sys.version_info >= (3, 10):
 else:
     from typing_extensions import Concatenate, ParamSpec, TypeAlias
 
-if sys.version_info >= (3, 8):
-    from typing import Literal
-else:
-    from typing_extensions import Literal
+from typing import Literal
 
 P = ParamSpec("P")
 T = TypeVar("T")
